Move more code from main.js to individual module files. Rename plugins -> modules. Remove some obsolete comments.
