+++ src/data/Thread.js
@@ -91,7 +91,6 @@ class Thread {
       dm_message_id: dmMessage.id
     });
 
-    // The string type check is due to a knex bug, see https://github.com/tgriesser/knex/issues/1276
     if (this.scheduled_close_at) {
       await this.cancelScheduledClose();
       await this.postSystemMessage(`Cancelling scheduled closing of this thread due to new reply`);
@@ -144,7 +143,6 @@ class Thread {
       dm_message_id: msg.id
     });
 
-    // The string type check is due to a knex bug, see https://github.com/tgriesser/knex/issues/1276
     if (this.scheduled_close_at) {
       await this.cancelScheduledClose();
       await this.postSystemMessage({
