+++ src/modules/snippets.js
